Node.js web application
A website built in Node.js to deploy cross platform games and web
apps.This project is created using Express and Socket.io to work on
cross platfrom such as for chatting or playing multiplayer games.

The main idea of the website to get an uderstanding of how to use
websockets and manipulate user data. The application uses the idea of
using multiple ports on one server to run individual application witout
cross passing data with each other.